backend/src/chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pypinyin
aiofiles
